+++ libs/application/templates/inheritance-report/src/fields/CalculationsOfTotal/CalculateTotalAssets/index.tsx
@@ -5,7 +5,7 @@ import { Box, Text } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
 import { FC, useEffect, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
-import { m } from '../../lib/messages'
+import { m } from '../../../lib/messages'
 
 export const CalculateTotalAssets: FC<FieldBaseProps> = ({ application }) => {
   const { answers } = application
