+++ libs/application/templates/inheritance-report/src/fields/index.ts
@@ -2,6 +2,6 @@ export { DoneImage } from './DoneImage'
 export { ReportFieldsRepeater } from './ReportFieldsRepeater'
 export { PrintScreen } from './PrintScreen'
 export { HeirsOverview } from './HeirsOverview'
-export { CalculateTotalAssets } from './CalculateTotalAssets'
-export { CalculateTotalDebts } from './CalculateTotalDebts'
-export { CalculateTotalBusiness } from './CalculateTotalBusiness'
+export { CalculateTotalAssets } from './CalculationsOfTotal/CalculateTotalAssets'
+export { CalculateTotalDebts } from './CalculationsOfTotal/CalculateTotalDebts'
+export { CalculateTotalBusiness } from './CalculationsOfTotal/CalculateTotalBusiness'
